+++ src/main/kotlin/app/morphe/gui/ui/theme/MorpheTypography.kt
@@ -1,3 +1,8 @@
+/*
+ * Copyright 2026 Morphe.
+ * https://github.com/MorpheApp/morphe-cli
+ */
+
 package app.morphe.gui.ui.theme
 
 import androidx.compose.runtime.Composable